ui: contract page mobile friendly (#144)
