+++ services/explorer-ui/src/pages/contract-instance-details/index.tsx
@@ -6,16 +6,9 @@ import { API_URL, aztecExplorer } from "~/service/constants";
 import { getContractData } from "./util";
 
 export const ContractInstanceDetails: FC = () => {
-  let address = "";
-  try {
-    const params = useParams({
-      from: "/contracts/instances/$address",
-    });
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    address = params.address;
-  } catch (e) {
-    // TODO
-  }
+  const { address } = useParams({
+    from: "/contracts/instances/$address",
+  });
   const {
     data: contractInstanceDetails,
     isLoading,
@@ -28,7 +21,7 @@ export const ContractInstanceDetails: FC = () => {
   if (!contractInstanceDetails) return <div>No data</div>;
 
   const apiEndpointUrl = `${API_URL}/${aztecExplorer.getL2ContractInstance(
-    address
+    address,
   )}`;
 
   return (
