+++ services/explorer-ui/src/routes/contracts/classes.$id.lazy.tsx
@@ -1,5 +1,5 @@
 import { createLazyFileRoute } from "@tanstack/react-router";
-import { ContractClassDetails } from "~/pages/contracts/class-details";
+import { ContractClassDetails } from "~/pages/contract-class-details";
 
 export const Route = createLazyFileRoute("/contracts/classes/$id")({
   component: ContractClassDetails,
