+++ services/explorer-ui/src/routes/contracts/instances.$address.lazy.tsx
@@ -1,5 +1,5 @@
 import { createLazyFileRoute } from "@tanstack/react-router";
-import { ContractInstanceDetails } from "~/pages/contracts/instance-details";
+import { ContractInstanceDetails } from "~/pages/contract-instance-details";
 
 export const Route = createLazyFileRoute("/contracts/instances/$address")({
   component: ContractInstanceDetails,
